Move project page components to app router
