+++ app/projects/card/social-media/styles.tsx
