+++ app/projects/card/styles.ts
